% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{add_rurality}
\alias{add_rurality}
\title{Merge Rurality Data onto a Data Frame}
\usage{
add_rurality(
  data,
  fips_col = "fips",
  vars = c("rurality_score", "rurality_classification", "rucc_2023")
)
}
\arguments{
\item{data}{A data frame with a FIPS code column.}

\item{fips_col}{The name of the FIPS code column (default: "fips").}

\item{vars}{Which rurality variables to add. Default adds score and
classification. Use "all" for all variables.}
}
\value{
The input data frame with rurality columns appended.
}
\description{
Joins rurality data onto an existing data frame by FIPS code.
}
\examples{
my_data <- data.frame(fips = c("05031", "06037", "48453"), value = 1:3)
add_rurality(my_data)
add_rurality(my_data, vars = "all")

}
