% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_rucc}
\alias{get_rucc}
\title{Get RUCC Code for a County}
\usage{
get_rucc(fips)
}
\arguments{
\item{fips}{A character vector of 5-digit county FIPS codes.}
}
\value{
An integer vector of RUCC codes (1-9), or NA for unmatched FIPS.
}
\description{
Returns the USDA Rural-Urban Continuum Code (2023) for one or more counties.
}
\examples{
get_rucc("05031")
get_rucc(c("05031", "06037"))

}
