% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{get_rurality}
\alias{get_rurality}
\title{Look Up Rurality Data by FIPS Code}
\usage{
get_rurality(fips)
}
\arguments{
\item{fips}{A character vector of 5-digit county FIPS codes.}
}
\value{
A tibble with rurality data for the matched counties.
}
\description{
Returns the full rurality record for one or more county FIPS codes.
}
\examples{
get_rurality("05031")
get_rurality(c("05031", "06037", "48453"))

}
