% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{rurality_score}
\alias{rurality_score}
\title{Get Rurality Score for a County}
\usage{
rurality_score(fips)
}
\arguments{
\item{fips}{A character vector of 5-digit county FIPS codes.}
}
\value{
A numeric vector of rurality scores, or NA for unmatched FIPS.
}
\description{
Returns the composite rurality score (0-100) for one or more counties.
}
\examples{
rurality_score("05031")
rurality_score(c("05031", "06037", "48453"))

}
